requirements.txt
cffi==1.15.1
crewai
composio-crewai
langchain-openai
python-dotenv
crewai-tools
streamlit
openai-whisper
yt_dlp